requirements.txt
flask==3.0.3
python-jose[cryptography]==3.3.0
requests==2.32.3
gunicorn==22.0.0
flask-cors==4.0.1
